Rename and reorganize classes and files.
Rename the Propeller2 class to P2Cog.
Add a class P2Hub which implements access to memory and ports.

Make the P2Hub class construct 1, 2, 4, 8, or 16 instances of COGs.
Give each P2Cog a pointer to its parent for Hub access.

The goal is to have the hardware structure of the Propeller2 modeled
in the class hierarchy. It should (later) be possible to use the Qt5
signal/slot mechanisms to implement events and interrupts between the
Hub and/or the Cogs. The Hub will then execute COGs round robin,
update counters and signal events exactly as the hardware does.