WIP: add disassembler view / model, first UI elements
Add a QAbstractTableModel derived class P2DasmModel to make an
easy to instantiate disassembler table view.

Clean up buggy code and comments in the P2Hub and P2Cog code to
access memory in the COGs, their LUTs, and the HUB memory.

Refactor disassembler to create the strings for the instruction
at PC in member variables, returned by ::dasm() as references.