WIP: Implement PAT / PIN / WAIT / LOCK states & flags
Implement some more of the state machinery required to actually
run COG instructions. Add more accessors to the P2Cog class.

Create a P2CogView class to show the current state of a COG.
Add a QTabWidget to the main window to switch between P2DasmView
and a view of the COGs with their states.