WIP: still refactoring large portions
The P2Token class now has a notion of token types (p2_tokentype_e).
This makes it easier for the rest of the code to handle
classes of tokens like operations (unary, binary, multiplications,
additions) and also the Propeller2 specific conditionals and
MODCZ parameters.

Improve the P2AsmModel to show info about errors and symbols (WIP).

Fix a lot of assembler functions which need to switch on the remaining
number of tokens on a line between the D,{#S} and D only variants,
or the D,{#S},#N and D only variants.

The P2Atom class is somewhat useful to keep data in variable sizes.
The extraction of multiple bytes, words, longs into opcodes, or
rather into data to put in memory, still needs work.