WIP: got the expression parsing working - almost
Parsing sub expressions inside parenthesis was a little difficult
to get right. Now it seems to be working.

The qz80 spin2 code shows some more issues left to fix in the
parsers, such as comments enclosed in curly braces later in a line.
And there's not only WC, WZ, and WCZ, but also "WC, WZ" or "WZ, WC".