WIP: Lots of refactoring again.
Things are getting better, but rather complicated at a first glance.
I hope to make this _the_ reference for assembling and disassembling
Propeller2 code, and also for emulation.

Spent a lot of time documenting the functions. The P2Token class now
has a lot more types defined to assist the lexing of source code,
and to also make the expression parsing follow definite rules.