WIP: Add a helper class P2AsmWords
The list of words and their token values are stored together
with the original source line's pos (start) and end of the
string that makes a word.

This can be used to extract or mark portions of the original
source string, i.e. source[line].mid(pos, end).

I think of making the P2AsmView use this info to show ToolTips
for the source column, depending on where the mouse cursor is.