WIP: Now trying hard to sort out traits
Make the P2Atom the carrier for a value's traits. For now, and until
I really understood everything, there are the following traits:

None		default
Immediate	expression started with '#'
Augmented	expression started with '##'
Relative	expression started with '@'
Absolute	expression started with '\'
AddressHub	expression started with '#@'
RelativeHub	expression started with '@@@'

Of course this means that the token for t__MOD is now '%', because
the backslash is the force absolute addressing trait.

Now that the atom carries the info, the index expressions in
square brackets after PTRA/PTRB can have their own trait without
changing that of the instruction.