WIP: P2Doc and new class P2DocOpcode overhaul
Trying to get the tricky beast that is Propeller2 opcode decoding
fit in a class which has the per opcode patter,
brief description, instruction example, flag effects description,
token, ...

Building a map of masks after all opcodes are set.
Building a list of one bits in the masks to scan the
hash with masks in drecreasign 1 bit density.

The uber-tricky TESTP/BITL etc. are still to fix.