WIP: Add class P2Union; refactor P2Atom; fixes
Add yet another class P2Union which is the class of the value
of a P2Atom. It can hold an arbitrary number of typed values
(TypedVar) as QVariantList.

It has functions to extract these typed lists into e.g.
QVector<p2_BYTE>, i.e. strings or byte array, words, longs, etc.

It is not yet complete in that reassembling what was defined
as multiple longs into an array of e.g. bytes is imperfect.

Also the other way round, combining (multiple) (lists of) bytes
into larger elements (words, longs, quads) is not yet done.

The P2Atom now is a wrapper around this class with the addional
information about the P2Atom type, which may differ from the
contained data types, and the traits, i.e. whether the P2Atom
is to be used as an immediate, HUB address, relative, forced augmentation,
etc.

Refactored a lot of other code and applied fixes.
The P2AsmModel is now capable of providing multi-instruction
opcodes to the QTableView for e.g. AUGD and AUGS prefixes.

It also displays multi longs data while in this part there
is work left to do.