WIP: improve P2Word / P2Token::tokenize() w/ QStringRef
The QStringRef is a light-weight helper class for QString. It already
has the position and length parameters which the P2Word class defined
on its own. Instead now keep the QStringRef in the P2Word and use
its m_ref->position() and m_ref->length() values where needed.

As a consequence the P2Token::tokenize() now needs a pointer to
the QString where the line is stored for the lifetime of the
P2Word objects. This is the QString in the m_source[0…count-1] list.

All this may seem awkward, but is supposed to improve the speed
because of the reduced overhead of creating (temporary) QStrings.