WIP: Excessive refactoring
The P2Union class needed an overhaul and is now derived from
a QVector<p2_uniont_t>.
The p2_union_t is a struct which stores the type and the
value of a single element, i.e. p2_BYTE, p2_WORD, p2_LONG,
p2_QUAD, p2_REAL, besides the standard bool and char.

The P2Union class is used as the P2Atoms storage. P2Atom now
does not only have a value, but also an (optional) index,
which is where an expression in square brackets is kept
for simplicity of the rest of the code.

That is a P2Atom can now store an expression like ptra++[4] or
even "Hello"[3*4].

The traits were moved from P2Atom::Traits enum to global name
space p2_traits_e. Also add some global bit set/clear/test functions
for p2_traits_e because casting enums to int and from int back to
enum just to add or remove a bit is cumbersome.

All in all things are beginning to get in shape.